Addressed reviews
